% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_load_data.R
\name{load_creditcard}
\alias{load_creditcard}
\title{Load Kaggle Credit Card Fraud Dataset}
\usage{
load_creditcard(path = NULL, sample_size = 10000, seed = 42)
}
\description{
Load Kaggle Credit Card Fraud Dataset
}
