% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_visualization.R
\name{plot_fraud_space}
\alias{plot_fraud_space}
\title{Plot PCA Fraud Space}
\usage{
plot_fraud_space(pca_result, em_result, max_points = 3000)
}
\arguments{
\item{pca_result}{List from \code{\link{pca_compress}}.}

\item{em_result}{List from \code{\link{em_fraud_model}}.}

\item{max_points}{Integer. Max points to plot (for speed). Default 3000.}
}
\value{
A ggplot2 object.
}
\description{
Scatter plot of transactions in PCA space (PC1 vs PC2), coloured by EM fraud
probability. Red = high fraud probability, green = low.
}
\examples{
\dontrun{
data  <- load_creditcard()
prep  <- preprocess_kaggle(data)
pca_r <- pca_compress(prep)
em_r  <- em_fraud_model(pca_r)
print(plot_fraud_space(pca_r, em_r))
}

}
