% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_train_test_predict.R
\name{print_test_results}
\alias{print_test_results}
\title{Print a Clean Test Results Summary}
\usage{
print_test_results(results)
}
\arguments{
\item{results}{A list returned by \code{\link{predict_on_test}}.}
}
\description{
Prints a formatted summary of test set performance after
\code{\link{predict_on_test}}.
}
\examples{
\dontrun{
data    <- load_creditcard()
splits  <- train_test_split(data)
trained <- train_model(splits$train)
results <- predict_on_test(trained, splits$test)
print_test_results(results)
}

}
